server/requirements.txt
anyio==3.6.1
cachetools==5.2.0
certifi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_884c889c-96af-444f-bd6d-daddb5e9a462ykj3l5n_/croots/recipe/certifi_1655968814730/work/certifi
charset-normalizer==2.1.1
click==8.1.3
fastapi==0.85.0
google-api-core==2.10.1
google-api-python-client==2.61.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.56.4
h11==0.13.0
httplib2==0.20.4
idna==3.4
numpy==1.23.3
oauthlib==3.2.1
opencv-python==4.6.0.66
pandas==1.4.4
Pillow==9.2.0
protobuf==4.21.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
pyzbar==0.1.9
qrcode==7.3.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.3.0
union==0.1.10
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.18.3
websocket-client==1.4.1
websockets==10.3
